Implemented click listener in Post component using UserContext
In this commit, I added a click listener to the Post component using UserContext. When a user clicks on a post, it captures the user uid from the click event and sends it to the Feed component via context. In the Feed component, filtered posts are displayed based on the uid of the user who clicked the post.