Implemented user click event in ConnectToUser component using UserContext for data filtering in Feed
Description:
In this commit, I  implemented a user click event in the  `ConnectToUser` component. When a user clicks on a profile image, the user ID (uid) is captured and sent to the `Feed` component using the `UserContext`. This change enables data filtering in the `Feed` component based on the clicked user's ID, allowing the display of filtered data based on the selected user.And the data is shown in the home page feed section .