Created & Added FollowUsers component to RightSidebar for user following functionality
**Description:**
In this commit, I Created the `FollowUsers` component and integrated it into the `RightSidebar`. This component is responsible for displaying a list of users. To ensure that the displayed data is different from what is fetched in the ConnectToUser component (which shares the same data source from Firebase), the list is shuffled using the Fisher-Yates shuffle algorithm. The "View More" button allows users to load additional users when available. This component sets the foundation for users to follow other users they are interested in!