Implemented currentUser click event in LeftSidebar component using UserContext for data filtering in Feed & resolved console errors
Description:
In this commit, I implemented a currentUser click event in the LeftSidebar component. This event only registers a click when the currentUser's profile image is visible. When a user clicks on their profile image in the LeftSidebar, the User ID (uid) of the currentUser is captured and sent to the Feed component via the UserContext. This change enables data filtering in the Feed component, allowing the display of currentUser data. Users can now view all their posts and manage them (delete, edit) if needed.
Additionally, resolved console errors  across all pages.